requirements.txt
pyTelegramBotAPI
python-dotenv
-e git+https://github.com/openai/whisper.git#egg=whisper